refactor: domain, view, controller 패키지 분리
