+++ src/main/java/Main.java
@@ -1,4 +1,4 @@
-import race.Race;
+import race.domain.Race;
 
 public class Main {
     public static void main(String[] args) {
