+++ src/main/java/race/domain/RaceInput.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 import java.util.List;
