+++ src/test/java/race/RacingCarTest.java
@@ -2,6 +2,7 @@ package race;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import race.domain.RacingCar;
 
 import static org.assertj.core.api.Assertions.assertThat;
